+++ src/components/admin/TableBody.tsx
@@ -3,49 +3,45 @@ import {
   TableCell,
   TableRow
 } from '@tremor/react';
-import { FaEdit, FaTrash } from 'react-icons/fa';
 
 import { Item, PriceHistory } from '@/server/models/item';
 
-export type ItemProps = {
+import { ActionIcons } from './Icons';
+
+export type TableBodyProps = {
   items: Item[];
-}
+};
 
-export type Mapping = {
-  item: string;
-  entry: string;
-  id: number;
-}
-
-const ActionIcons = () => (
-  <div className="flex items-center">
-    <FaEdit className="mr-3 cursor-pointer" title="Edytuj" />
-    <FaTrash className="mr-3 cursor-pointer" title="Usuń" />
-  </div>
-);
-
-export const renderPriceHistory = (priceHistory: PriceHistory[]) =>
+const renderPriceHistory = (priceHistory: PriceHistory[]) =>
   priceHistory.map((entry, id) => (
     <div key={id}>
       {' '}
-      {entry.date.toLocaleDateString()}: {entry.price}
+      {entry.date.toLocaleDateString()}: {entry.price} zł
     </div>
   ));
 
-const TableBody = ({ items }: ItemProps) => {
-  const renderTableCells = (item: Item) =>
-    Object.values(item).map((value, index) => (
-      <TableCell key={index}>
-        {value === 'priceHistory' ? renderPriceHistory(value) : value}
-      </TableCell>
-    ));
-
+const TableBody = ({ items }: TableBodyProps) => {
   return (
-    <TremorTableBody className="border">
+    <TremorTableBody>
       {items ? (
         items.map((item: Item, id) => (
           <TableRow key={id}>
-            {renderTableCells(item)}
+            <TableCell>{item.name}</TableCell>
+            <TableCell>{item.description}</TableCell>
+            <TableCell>{item.price} zł</TableCell>
+            <TableCell>{item.quantity}</TableCell>
+            <TableCell>
+              <div
+                className={`pb-0.75 pt-0.75 w-12 pl-3 ${item.visible ? 'bg-green-400' : 'bg-red-400'} rounded-tremor-small text-white`}
+              >
+                {item.visible}
+              </div>
+            </TableCell>
+            <TableCell>{item.category}</TableCell>
+            <TableCell>
+              {renderPriceHistory(item.priceHistory as PriceHistory[])}
+            </TableCell>
+            <TableCell>{item.photos}</TableCell>
             <TableCell>
               <ActionIcons />
             </TableCell>
