--- src/components/dashboard/TableHead.tsx
@@ -1,19 +1,0 @@
-import {
-  TableHead as TremorTableHead,
-  TableHeaderCell,
-  TableRow
-} from '@tremor/react';
-
-import { Title } from '../utils/constants';
-
-const TableHead = () => (
-  <TremorTableHead className="bg-tremor-background-subtle">
-    <TableRow>
-      {Object.values(Title).map(item => (
-        <TableHeaderCell key={item}>{item}</TableHeaderCell>
-      ))}
-    </TableRow>
-  </TremorTableHead>
-);
-
-export default TableHead;
