--- src/components/dashboard/index.ts
@@ -1,5 +1,0 @@
-import ButtonProduct from './ButtonProduct';
-import TableBody from './TableBody';
-import TableHead from './TableHead';
-
-export { ButtonProduct, TableBody, TableHead };
